requirement.txt
fastembed
qdrant-client
langchain_community
langchain
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
exceptiongroup==1.2.0
fastembed==0.2.2
filelock==3.13.1
flatbuffers==24.3.7
fsspec==2024.3.0
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
loguru==0.7.2
mpmath==1.3.0
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.17.1
packaging==24.0
portalocker==2.8.2
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
PyYAML==6.0.1
qdrant-client==1.8.0
requests==2.31.0
sniffio==1.3.1
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
elasticsearch
sentence_transformers
aiogram==3.6.0
python-igraph
gravis

